Loop counter in PascalRectangle and a test in CoinsTest
Changed loop counter to long in PascalRectangle and added a test to
Coins

+++ javaExercises/src/main/java/com/capgemini/pascalrectangle/PascalRectangle.java
@@ -30,7 +30,7 @@ public class PascalRectangle {
 
 		// calculate (r - c)!
 		long rmcfactorial = 1;
-		for (int i = 1; i <= (r - c); i++) {
+		for (long i = 1; i <= (r - c); i++) {
 			rmcfactorial *= i;
 		}
